% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Get Linear Regression Coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{linreg}}

\item{...}{additional arguments}
}
\value{
a vector containing the regression coefficients
}
\description{
An S3 function to return coefficients from a \code{linreg} object
}
