% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_ridgereg.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Get Ridge Regression Coefficients}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ridgereg}}

\item{...}{additional arguments}
}
\value{
a vector containing the regression coefficients
}
\description{
An S3 function to return coefficients from a \code{ridgereg} object
}
