% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix_data_frame.R
\name{model.matrix.data.frame}
\alias{model.matrix.data.frame}
\title{Get Design Matrix}
\usage{
\method{model.matrix}{data.frame}(X, factor_indep_vars, add_intercept)
}
\arguments{
\item{X}{a \code{data.frame}}

\item{factor_indep_vars}{a vector of names of independent variables whose classes are of \code{factor} type}

\item{add_intercept}{a boolean; TRUE requires the design matrix to include a column for an intercept}
}
\value{
the design matrix
}
\description{
An S3 function to return a design matrix given a \code{data.frame} and
a linear regression \code{formula}
}
