% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix_data_frame.R
\name{model.matrix}
\alias{model.matrix}
\title{S3 Generic function - rewritten}
\usage{
model.matrix(X, factor_indep_vars, add_intercept)
}
\arguments{
\item{X}{an object}

\item{factor_indep_vars}{an object}

\item{add_intercept}{an object}
}
\description{
S3 Generic function - rewritten
}
\seealso{
\link{model.matrix.data.frame}
}
