% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{new_linreg}
\alias{new_linreg}
\title{Linreg Class Constructor}
\usage{
new_linreg(x = list())
}
\arguments{
\item{x}{A list}
}
\value{
The same list input with class attribute modified to 'linreg'
}
\description{
Linreg Class Constructor
}
