% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{new_ridgereg}
\alias{new_ridgereg}
\title{Ridgereg Class Constructor}
\usage{
new_ridgereg(x = list())
}
\arguments{
\item{x}{A list}
}
\value{
The same list input with class attribute modified to 'ridgereg'
}
\description{
Ridgereg Class Constructor
}
