% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plots of Linear Regression Residuals}
\usage{
\method{plot}{linreg}(x, num_labels = 3, labels_by_val = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{linreg}}

\item{num_labels}{an integer, specifying the number of extreme residuals to label}

\item{labels_by_val}{a boolean; TRUE forces the function to label selected extreme points by value}

\item{...}{additional arguments}
}
\value{
a vector containing the regression coefficients
}
\description{
An S3 function to return plots from a \code{linreg} object
}
