% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Return Modelled Response Values (predictions given training data)}
\usage{
\method{pred}{linreg}(x)
}
\arguments{
\item{x}{an object of class \code{linreg}}
}
\value{
a vector containing the modelled response values
}
\description{
An S3 function to return modelled response values from a \code{linreg} object
}
