% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict Response using \code{ridgereg} Model}
\usage{
\method{predict}{ridgereg}(ridgereg_obj, data = NULL)
}
\arguments{
\item{ridgereg_obj}{an object of class \code{ridgereg}}

\item{data}{set to NULL by default, returning values fitted by the model on the training dataset;
when a non-trivial dataset is provided as input that is similar to the training dataset, the function
returns predicted response values for the input}
}
\value{
a vector containing the modelled response values
}
\description{
An S3 function to return modelled response values from observations of 
independent variables using a \code{ridgereg} model
}
