% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print Out \code{linreg} Function Call and Regression Coefficients}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{an object of class \code{linreg}}

\item{...}{additional parameters}
}
\description{
Print Out \code{linreg} Function Call and Regression Coefficients
}
