% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print Out \code{ridgereg} Function Call and Coefficients}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ridgereg}}

\item{...}{additional parameters}
}
\description{
Print Out \code{ridgereg} Function Call and Coefficients
}
