% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Return Linear Regression Residuals}
\usage{
\method{resid}{linreg}(x)
}
\arguments{
\item{x}{an object of class \code{linreg}}
}
\value{
a vector containing the residuals
}
\description{
An S3 function to return residuals from a \code{linreg} object
}
