% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression}
\usage{
ridgereg(formula, data, lambda, qr_method = FALSE, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. The details of model specification are
given under ‘Details’.}

\item{data}{a data frame containing the variables in the model}

\item{lambda}{hyperparameter for ridge regression}

\item{qr_method}{a boolean; TRUE forces ridgereg to use QR decomposition for the regression}

\item{...}{additional arguments to be passed to the \code{model.matrix} function, see 'Details'}
}
\value{
the function returns an object of class "ridgereg", which has a number of associated 
methods including \code{\link[linear.regression]{print.ridgereg}}, 
\code{\link[linear.regression]{predict.ridgereg}},
and \code{\link[linear.regression]{coef.ridgereg}}.
The "ridgereg" object itself is a list containing at least the following components:
\item{call}{the matched call}
\item{coefficients}{a named vector of coefficients}
\item{fitted_values}{the fitted mean values}
\item{residuals}{the residuals, that is response minus fitted values}
\item{df}{the residual degrees of freedom}
\item{residual_variance}{estimated value of the variance of the residuals}
\item{t_values}{t-statistics of the independent variables, that is coefficients/standard-error}
\item{p_values}{p-values for the two-sided t-test with null that coefficients are each zero}
\item{std_means}{a vector of means of the independent variables in the training dataset}
\item{std_sd}{a vector of standard deviations of the independent variables in the training dataset}
}
\description{
The 'ridgereg' function runs a ridge regression algorithm on a collection of 
observations of a dependent variable and corresponding independent variables. Thereafter, it
provides views of regression coefficients, residuals, and other statistics
}
\details{
Models for \code{linreg} are specified symbolically. A typical model has the form \code{response ~ terms} or
\code{response ~ .}, where 'response' is the (numeric) response vector. In the former form, 'terms' is a series
of terms which specifies a linear predictor for response. In the latter form, the '.' indicates that all available
predictors must be taken into account for modelling. Term specification of the form \code{first + second}
indicates all the terms in first together with all the terms in second with duplicate terms removed.
Specifications of the form \code{first:second} or \code{first*second} are recognised exactly the same way as 
\code{first+second}. Additional arguments passed via ..., if provided, will force the linear model will 
have no intercept.
}
\note{
The 'ridgereg' function was designed to work like the 'lm.ridge' function in the 'MASS' 
package. Independent variables are standardised before regression is performed; the predict method
on the returned object will use the same means and standard deviations used for standardisation of
the training dataset. Review the 'See Also' section.
}
\examples{
library(mlbench)
data(BostonHousing)
ridgereg(medv ~ crim + zn + indus + chas + nox + rm + age + dis + rad, BostonHousing, lambda = 1)
}
\seealso{
\code{\link[MASS]{lm.ridge}}

\code{\link[stats]{lm}}
}
