% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summarise Linear Regression output}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{linreg}}

\item{...}{additional parameters}
}
\description{
Summarise Linear Regression output
}
